backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
python-jose[cryptography]
